requirements.txt
python-gitlab~=2.7.1

PySimpleGUI~=4.43.0
PyYAML~=5.4.1
